requirements.txt
autoflake==2.2.0
python-dotenv
deeplake==3.6.19
langchain==0.0.266
openai==0.27.8
pypdf==3.15.1
tiktoken==0.4.0
jq==1.4.1
requests==2.31.0

